Refactored StoryViewer component, and fixed some styling
