+++ package.json
@@ -25,6 +25,7 @@
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
+    "react-sizeme": "^2.6.12",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0"
   },
