+++ src/styles/less/StoryViewer.less
@@ -14,3 +14,9 @@
     width: 20%;
   }
 }
+
+.page-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
